Fix code health issues and refactor timing utilities
- Refactored custom throttle/debounce logic into a centralized utility in src/lib/timing.ts.
- Simplified scroll throttle cleanup in Navigation.tsx.
- Imported DependencyList from react in useHotkeys.ts.
- Wrapped localStorage access in try/catch in useBlogDrafter.ts.
- Added error handling and UI feedback for WCSScraperTool data fetching.
- Optimized useUXAuditor.ts to skip network requests if apiKey is missing.
- Fixed JSX tag mismatch and normalized indentation in GlobalSearch.tsx.
- Guarded import.meta.glob in content.ts to support SSR/test environments.
- Updated Playwright visual regression snapshots.

All unit tests and linters pass. Audit tool shows no anti-patterns.
