--- src/hooks/utils/useDebounce.ts
@@ -1,20 +1,0 @@
-import { useState, useEffect } from 'react';
-
-/**
- * Hook for debouncing a value.
- */
-export function useDebounce<T>(value: T, delay: number): T {
-  const [debouncedValue, setDebouncedValue] = useState<T>(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
