+++ src/lib/timing.ts
@@ -1,0 +1,34 @@
+/**
+ * Simple throttle implementation.
+ */
+export function throttle<Args extends unknown[]>(
+  func: (...args: Args) => void,
+  limit: number
+): (...args: Args) => void {
+  let inThrottle = false;
+  return (...args: Args) => {
+    if (!inThrottle) {
+      func(...args);
+      inThrottle = true;
+      setTimeout(() => {
+        inThrottle = false;
+      }, limit);
+    }
+  };
+}
+
+/**
+ * Simple debounce implementation.
+ */
+export function debounce<Args extends unknown[]>(
+  func: (...args: Args) => void,
+  wait: number
+): (...args: Args) => void {
+  let timeout: ReturnType<typeof setTimeout> | null = null;
+  return (...args: Args) => {
+    if (timeout) {
+      clearTimeout(timeout);
+    }
+    timeout = setTimeout(() => func(...args), wait);
+  };
+}
